feat: Rename MinaSDK.ts to MinaSigner.ts
