+++ frontend/mina-signer/package.json
@@ -15,12 +15,12 @@
     "cryptocurrency"
   ],
   "author": "o1labs",
-  "license": "MIT",
+  "license": "Apache-2.0",
   "homepage": "https://minaprotocol.com/",
   "repository": "https://github.com/MinaProtocol/mina",
   "bugs": "https://github.com/MinaProtocol/mina/issues",
-  "main": "dist/src/MinaSDK.js",
-  "types": "dist/src/MinaSDK.d.ts",
+  "main": "dist/src/MinaSigner.js",
+  "types": "dist/src/MinaSigner.d.ts",
   "files": [
     "dist"
   ],
