+++ frontend/mina-signer/tests/keypair.test.ts
@@ -1,4 +1,4 @@
-import Client from "../src/MinaSDK";
+import Client from "../src/MinaSigner";
 
 describe("Keypair", () => {
   let client: Client;
